+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## [unreleased]
+## [0.10.0]
 
 ### Added
 * Add support for new Diaspora protocol ISO 8601 timestamp format introduced in protocol version 0.1.6.
+* Tests are now executed also against Python 3.6.
 
 ### Fixes
 * Don't crash `federation.utils.diaspora.retrieve_diaspora_webfinger` if XRD parse raises an `xml.parsers.expat.ExpatError`.